#1: Separated ElevTile class into separate files. Fixed bug iniimage scaling of elevation subtiles.
