+++ qt/tileedit/tileedit/tile.cpp
@@ -1,6 +1,5 @@
 #include "tile.h"
 #include "ddsread.h"
-#include "cmap.h"
 #include <iostream>
 #include <algorithm>
 
@@ -117,124 +116,4 @@ std::pair<MaskTile*,NightlightTile*> MaskTile::Load(const std::string &root, int
 		ltile->img.data[i] |= 0xFF000000;
 	}
 	return std::make_pair(mtile, ltile);
-}
-
-
-ElevTile::ElevTile(int lvl, int ilat, int ilng)
-	: Tile(lvl, ilat, ilng)
-{
-	lat_subrange.second = 256;
-	lng_subrange.second = 256;
-}
-
-ElevTile::ElevTile(const ElevTile &etile)
-	: Tile(etile)
-	, m_edata(etile.m_edata)
-{
-}
-
-double ElevTile::nodeElevation(int ndx, int ndy)
-{
-	int idx = (ndy + 1)*m_edata.width + (ndx + 1);
-	return (double)m_edata.data[idx];
-}
-
-bool ElevTile::Load(const std::string &root)
-{
-	// read tile from Elev layer
-	char path[1024];
-	sprintf(path, "%s/%s/%02d/%06d/%06d.elv", root.c_str(), Layer().c_str(), m_lvl, m_ilat, m_ilng);
-	m_edataBase = elvread(path);
-	m_edata = m_edataBase;
-
-	if (m_edata.data.size()) {
-		// read modifications from Elev_mod layer if present
-		sprintf(path, "%s/%s_mod/%02d/%06d/%06d.elv", root.c_str(), Layer().c_str(), m_lvl, m_ilat, m_ilng);
-		elvmodread(path, m_edata);
-	} else {
-		// interpolate from ancestor
-		LoadSubset(root, this);
-	}
-
-	// turn elevation data into an image
-	if (m_edata.data.size())
-		ExtractImage();
-
-	return m_edata.data.size() > 0;
-}
-
-void ElevTile::LoadSubset(const std::string &root, ElevTile *tile)
-{
-	if (tile->m_sublvl > 1) {
-		tile->lat_subrange.first /= 2;
-		tile->lat_subrange.second /= 2;
-		tile->lng_subrange.first /= 2;
-		tile->lng_subrange.second /= 2;
-		if (tile->m_subilat & 1) {
-			tile->lat_subrange.first += 128;
-			tile->lat_subrange.second += 128;
-		}
-		if (tile->m_subilng & 1) {
-			tile->lng_subrange.first += 128;
-			tile->lng_subrange.second += 128;
-		}
-		tile->m_sublvl -= 1;
-		tile->m_subilat /= 2;
-		tile->m_subilng /= 2;
-
-		char path[1024];
-		sprintf(path, "%s/%s/%02d/%06d/%06d.elv", root.c_str(), Layer().c_str(), tile->m_sublvl, tile->m_subilat, tile->m_subilng);
-		tile->m_edata = elvread(path);
-		if (tile->m_edata.data.size()) {
-			sprintf(path, "%s/%s_mod/%02d/%06d/%06d.elv", root.c_str(), Layer().c_str(), tile->m_sublvl, tile->m_subilat, tile->m_subilng);
-			elvmodread(path, tile->m_edata);
-			tile->m_edata = tile->m_edata.SubTile(tile->lng_subrange, tile->lat_subrange);
-		}
-		else {
-			LoadSubset(root, this);
-		}
-	}
-}
-
-ElevTile *ElevTile::Load(const std::string &root, int lvl, int ilat, int ilng)
-{
-	ElevTile *etile = new ElevTile(lvl, ilat, ilng);
-	etile->Load(root);
-	return etile;
-}
-
-void ElevTile::ExtractImage(int exmin, int exmax, int eymin, int eymax)
-{
-	img.width = (m_edata.width-2)*2 - 2;
-	img.height = (m_edata.height-2)*2 - 2;
-	img.data.resize(img.width * img.height);
-
-	int dmin = (int)*std::min_element(m_edata.data.begin(), m_edata.data.end());
-	int dmax = (int)*std::max_element(m_edata.data.begin(), m_edata.data.end());
-
-	const Cmap &cm = cmap(CMAP_TOPO);
-
-	int imin = (exmin < 0 ? 0 : max(0, (exmin - 1) * 2 - 1));
-	int imax = (exmax < 0 ? img.width : min((int)img.width, exmax * 2));
-	int jmin = (eymax < 0 ? 0 : max(0, (int)img.height - (eymax - 1) * 2));
-	int jmax = (eymin < 0 ? img.height : min((int)img.height, (int)img.height - (eymin - 1) * 2 + 1));
-
-	for (int j = jmin; j < jmax; j++)
-		for (int i = imin; i < imax; i++) {
-			int ex = (i + 1) / 2 + 1;
-			int ey = (img.height - j) / 2 + 1;
-			INT16 d = m_edata.data[ex + ey * m_edata.width];
-			//INT16 d = edata.data[(img.height - j) * edata.width + i + 1];
-			int v = min((((int)d - dmin) * 256) / (dmax - dmin), 255);
-			if (v < 0 || v > 255)
-				std::cerr << "Problem" << std::endl;
-			img.data[i + j * img.width] = (0xff000000 | cm[v]);
-				//(0xFF << 0x18) | // alpha
-				//(v << 0x10) | // red
-				//(v << 0x08) | // green
-				//v; // blue
-
-			if (img.data[i + j * img.width] != (0xff000000 | cm[v]))
-				std::cerr << "Problem" << std::endl;
-		}
 }
