+++ qt/tileedit/tileedit/tile.h
@@ -4,7 +4,6 @@
 #include <windows.h>
 #include <vector>
 #include "ddsread.h"
-#include "elvread.h"
 
 class Tile
 {
@@ -73,28 +72,6 @@ public:
 
 protected:
 	MaskTile(int lvl, int ilat, int ilng);
-};
-
-class ElevTile : public Tile
-{
-public:
-	ElevTile(const ElevTile &etile);
-	static ElevTile *Load(const std::string &root, int lvl, int ilat, int ilng);
-	const std::string Layer() const { return std::string("Elev"); }
-	double nodeElevation(int ndx, int ndy);
-	ElevData &getData() { return m_edata; }
-	ElevData &getBaseData() { return m_edataBase; }
-	void dataChanged(int exmin = -1, int exmax = -1, int eymin = -1, int eymax = -1) { ExtractImage(exmin, exmax, eymin, eymax); }
-
-protected:
-	ElevTile(int lvl, int ilat, int ilng);
-	bool Load(const std::string &root);
-	void LoadSubset(const std::string &root, ElevTile *tile);
-	void ExtractImage(int exmin = -1, int exmax = -1, int eymin = -1, int eymax = -1);
-
-private:
-	ElevData m_edata;
-	ElevData m_edataBase;
 };
 
 #endif // TILE_H
